feat(univ.cache.ts): add UnivCache class to handle caching of university data
The UnivCache class is added to handle caching of university data. It extends the BaseCache class from the redis.base module and provides methods to add university data to the cache and retrieve university data from the cache.

The addUnivToCache method is used to add university credentials to the cache. It connects to the Redis client, sets the university data using the HSET command, and creates a new job to add the new record to the database.

The getUnivFromCache method is used to retrieve university data from the cache. It connects to the Redis client, retrieves the data using the hGet command, and deserializes the data before returning it.

The serializeData method is used to serialize the university data before storing it in the cache.

The deserializeData method is used to deserialize the university data retrieved from the cache.

This class is exported as univCache for use in other modules.