feat(univ.queue.ts): add new file univ.queue.ts to handle university related jobs in a queue
The univ.queue.ts file is added to the project to handle university related jobs in a queue. It imports the BaseQueue class from the '@helpers/queue/base.queue' module and the univWorker from the './univ.worker' module. It also imports the Job interface from the 'bullmq' module.

The UnivQueue class extends the BaseQueue class and takes the queueWorkerName as a parameter in the constructor. It calls the super() method to initialize the base queue with the queueWorkerName and the univWorker.addNewJob function.

The class also defines an async method addUnivJob() which takes jobName and job as parameters. It calls the addJobToQueue() method from the base queue to add the job to the queue.

Finally, the univQueue object is exported as a new instance of the UnivQueue class with the queueWorkerName set to 'univQueue'.