feat(univ.services.ts): add UnivServices class with create and getUnivByRoot methods
This commit adds a new file `univ.services.ts` which contains the `UnivServices` class. The `UnivServices` class has two methods: `create` and `getUnivByRoot`.

The `create` method takes in a `univData` parameter of type `CreateUniversityRequest` and creates a new university record in the database using the `University.create` method. If an error occurs during the creation process, it is logged using the `log.error` method.

The `getUnivByRoot` method takes in a `root` parameter of type `string` and retrieves a university record from the database using the `University.findOne` method.

This commit adds the necessary functionality to interact with the university model in the application.