feat(department.cache.ts): add DepartmentCache class to handle caching of department data
The DepartmentCache class is responsible for caching department data. It extends the BaseCache class from the redis.base module. The constructor initializes the cache with the name 'departmentCache'.

The class provides a method addDepartmentToCache() to add department data to the cache. If the cache client is not open, it connects to the client. Then, it uses the HSET command to set the department data in the cache using the departmentKey() function from the redis.keys module.

The class also provides two helper methods, serializeData() and deserializeData(), to serialize and deserialize the department data respectively.

This commit adds the DepartmentCache class to handle caching of department data.