feat(department.queue.ts): add DepartmentQueue class to handle department-related jobs in the queue
The DepartmentQueue class is responsible for managing department-related jobs in the queue. It extends the BaseQueue class from the @helpers/queue module and uses the departmentWorker.addNewJob function as the worker for processing the jobs.

The DepartmentQueue class has a constructor that takes a queueWorkerName parameter and passes it to the super() method of the BaseQueue class. It also defines an addDepartmentJob method that adds a job to the queue using the addJobToQueue method inherited from the BaseQueue class.

The departmentQueue constant is an instance of the DepartmentQueue class, initialized with the queueWorkerName 'departmentQueue'. This constant can be used to interact with the department-related jobs in the queue.